defaults.mk: Add -Wno-nonnull-compare to CPPFLAGS
This warning is the dumbest misfeature I've ever seen.  It complains if
you check for NULL on a __nonnull__ parameter, completely ignoring that
the compiler generating the calls might not know about __nonnull__.

What it *should* do is annotate the landing sites during compilation,
and then during the link phase, check if it's exposed or not.  If the
function is static or hidden, it should give the warning that it does
now.  If it's an external symbol, it should throw a warning if you
*don't* check the arguments for NULL.

Signed-off-by: Peter Jones <pjones@redhat.com>